% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Coordinates}
\alias{Coordinates}
\title{Coordinates Testdata}
\format{
A data.frame
}
\source{
NA
}
\usage{
Coordinates
}
\description{
Coordinates lon and lat for two locations: DWD and Aukland
}
\keyword{datasets}
