% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat2UTM.R
\name{lonlat2UTM}
\alias{lonlat2UTM}
\title{Convert lon and lat into integer}
\usage{
lonlat2UTM(lonlat)
}
\arguments{
\item{lonlat}{Numerical vector with lon and lat, like c(50.10267, 8.74782)}
}
\value{
integer
}
\description{
Converts longitute and lattitude values to an integer to convert further into UTM
}
\examples{
lonlat2UTM(c(50.10267, 8.74782))
}
