% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MapColors}
\alias{MapColors}
\title{Five colors for maps}
\format{
A vector with 5 elements:
}
\source{
\url{https://colorbrewer2.org/#type=sequential&scheme=OrRd&n=5}
}
\usage{
MapColors
}
\description{
The dataset contains a vector with five colors for maps
Source: colorbrewer2.org
}
\keyword{datasets}
